signs: text size in blocks; auto-sized signs grow up from a fixed base
`size` was an opaque font-pixel value divided by the sign width (a fixed
1200px viewport constant leaked into the API). It's now the text height
in blocks — the same unit as width/height — so a sign is width×height
blocks with text `size` blocks tall. Convert the engine, migrate the
tutorial and course signs and the sign-menu docs (old ÷ 1200), and keep
the defaults byte-for-byte identical (250/1200).

Auto-height signs (bots, mainly) now anchor their bottom edge and grow
upward, so a long message climbs into the air instead of clipping down
through the speaker's head.

Retune the course welcome and progress signs to fill their faces.

+++ src/controllers/script_controllers/host_bridge.nim
@@ -924,7 +924,7 @@ proc new_markdown_sign(
     more: string,
     width: float,
     height: float,
-    size: int,
+    size: float,
     billboard: bool,
 ): Thing =
   result = Sign.init(
@@ -949,7 +949,7 @@ proc update_markdown_sign(
     more: string,
     width: float,
     height: float,
-    size: int,
+    size: float,
     billboard: bool,
 ) =
   thing.width = width
@@ -967,7 +967,7 @@ proc `height=`(self: Sign, value: float) =
   types.`height=`(self, value)
   self.message_value.touch self.message
 
-proc `size=`(self: Sign, value: int) =
+proc `size=`(self: Sign, value: float) =
   types.`size=`(self, value)
   self.message_value.touch self.message
